prevent deletion of employees and vehicles with active records

+++ src/Application/Vehicles/Commands/DeleteVehicle/DeleteVehicleCommand.cs
@@ -1,10 +1,12 @@
-﻿using System.Threading;
+﻿using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using CarsManager.Application.Common.Exceptions;
 using CarsManager.Application.Common.Interfaces;
 using CarsManager.Application.Common.Security;
 using CarsManager.Domain.Entities;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 
 namespace CarsManager.Application.Vehicles.Commands.DeleteVehicle
 {
@@ -28,10 +30,15 @@ namespace CarsManager.Application.Vehicles.Commands.DeleteVehicle
 
         public async Task<Unit> Handle(DeleteVehicleCommand request, CancellationToken cancellationToken)
         {
-            var entity = await context.Vehicles.FindAsync(request.Id);
+            var entity = await context.Vehicles
+                .Include(v => v.RoadBookEntries)
+                .FirstOrDefaultAsync(v => v.Id == request.Id);
 
             if (entity == null)
                 throw new NotFoundException(nameof(Vehicle), request.Id);
+
+            if (entity.RoadBookEntries.Any(r => r.ActiveUsers.Count > 0))
+                throw new ForbiddenVehicleDeletionException($"Vehicle {entity.LicencePlate} has assigned employees.");
 
             if (!string.IsNullOrEmpty(entity.Image))
                 imageManager.DeleteFile(request.ImagePath, entity.Image);
